pyproject.toml
[project]
name = "council-of-maniacs"
version = "0.1.0"
description = "A deliberation council of wildly unhinged AI personas"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
]
